+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Set up LSP client without `nvim-lspconfig` (removes the dependency).
+### Added
+- `HlsStart`, `HlsStop` and `HlsRestart` commands.
+- Dynamically load `haskell-language-server` settings JSON from project root, if available.
 
 ## [1.7.0] - 2023-01-27
 ### Fixed